Adapt the vectortrend tests to components
